Refactored Edit component
